try fix some documentation warnings

+++ cocos/base/CCVector.h
@@ -194,7 +194,7 @@ public:
     
     /** 
      * Requests that the vector capacity be at least enough to contain n elements.
-     * @param capacity Minimum capacity requested of the Vector.
+     * @param n - The minimum capacity requested of the Vector.
      */
     void reserve(ssize_t n)
     {
@@ -217,7 +217,7 @@ public:
      */
     ssize_t size() const
     {
-        return  _data.size();
+        return  static_cast<ssize_t>(_data.size());
     }
     
     /** @brief Returns whether the Vector is empty (i.e. whether its size is 0).
@@ -268,7 +268,7 @@ public:
     T at(ssize_t index) const
     {
         CCASSERT( index >= 0 && index < size(), "index out of range in getObjectAtIndex()");
-        return _data[index];
+        return _data[static_cast<size_t>(index)];
     }
 
     /** Returns the first element in the Vector. */
